Soft fail and minimum time for busy wait for #132
Allow busy-wait to wait for a minimum sensible time and timeout without raising a RuntimeError.

Additionally force busy-wait to wait for at least the given timeout if the pin is high.

This should - hopefully - correct issues with some Impressions panels sticking perpetually in a busy wait.

Warnings will be emitted in both "failure" cases.